Properly detach process
Detached processes could be killed when the launching terminal closes.

Gnome terminal reproduction:
1. Run gnome-terminal -- opn file /path/to/file
2. Select program and open detached
3. Program will never be seen because it is instantly killed

In case of foot, reproduction is less easy because it seems to detach
processes automatically though it is timing dependent.
If there is just a small delay between the program being started and
opn exiting, the program survives. This small delay can be as simple
as running a build of opn with the -race flag.